Restructure some code and begin making classes for character data
The method 'getCharacterHash()' has been removed, a hash will always be generated. This reduces the code in the API call.

Fixed an issue in toArray() when an array type exists, it needs to loop through the values and all toArray on each one (An array is just an array of other objects)

Moved parse code into traits.

Moved city/gc/guardian back a folder, this folder might get quite long but most of them are single instances (collectable, classjob, item, etc)

Set character id for profile during construct, make it a requirement.

Added "initialize" to parse helper which does the usual: ensureHtml, trim html, setInitialDocument. Nearly all parse methods does this proceedure.
